rename SecurityPage to SecurityTab
