+++ app/components/views/PrivacyPage/PrivacyTab/ConfigMixer/CreateNeededAccounts.jsx
@@ -3,7 +3,7 @@ import { PassphraseModalButton } from "buttons";
 import { WatchOnlyWarnNotification } from "shared";
 import { AddMixerAccountsModal } from "modals";
 import { MIXED_ACCOUNT, CHANGE_ACCOUNT } from "constants";
-import style from "../Privacy.module.css";
+import style from "../PrivacyTab.module.css";
 
 const CreateNeededAccounts = ({
   onSubmit,
