--- app/components/views/PrivacyPage/SecurityPage/index.js
@@ -1,1 +1,0 @@
-export { default } from "./SecurityPage";
