+++ app/components/views/PrivacyPage/SecurityTab/SignMessage/SignMessage.jsx
