+++ app/components/views/PrivacyPage/SecurityTab/SignMessage/Signature.jsx
