+++ app/components/views/PrivacyPage/SecurityTab/SignMessage/hooks.js
