+++ app/components/views/PrivacyPage/SecurityTab/SignMessage/index.js
