+++ app/components/views/PrivacyPage/SecurityTab/ValidateAddress/ValidateAddress.jsx
