+++ app/components/views/PrivacyPage/SecurityTab/ValidateAddress/hooks.js
