+++ app/components/views/PrivacyPage/SecurityTab/ValidateAddress/index.js
