+++ app/components/views/PrivacyPage/SecurityTab/VerifyMessage/VerifyMessage.jsx
