+++ app/components/views/PrivacyPage/SecurityTab/VerifyMessage/hooks.js
