+++ app/components/views/PrivacyPage/SecurityTab/VerifyMessage/index.js
