+++ app/components/views/PrivacyPage/SecurityTab/index.js
@@ -1,0 +1,1 @@
+export { default } from "./SecurityTab";
